SocialNetworkHTML v2.0
Fixed display of avatars on the message page. Messages on the wall are randomly generated. Did an avatar pop up on hover. On the settings page, when you click on any item - it can be edited. Check for new passwords matching (if they do not match, the form will not be submitted). TODO: user avatars in the friends menu. User avatars and nicknames on the wall. When explicitly calling forms.submit (), the onsubmit method is ignored (I don’t know how to fix it, the form is sent anyway)